Moved SeqList and SeqCheckpoint into mod flow
